Strip bgworker package from test stuff
The way this package was written was a sort of mix of the
background_process lib and an example. Now that we are aiming to
shipping this as a simple lib package we strip away the test stuff since
anyone depending on and importing this won't be interested in importing
tests.