Fix stop check
Remove the double invocation of the test-stop target. That wasn't
actually a mistake. I wanted to invoke it twice just to be certain it
had stopped and wasn't slow or anything. I've compensated somewhat by
increasing the sleep between the two snapshots of show output we grab
for the comparison.