Avoid false positives
We wait for 2 seconds in the tests between fetching the counter value and then
determine if they have changed, i.e. if the background worker is running or not.
Sometimes we get a false positive as the counter value is the same. By lowering
the sleep in the background worker we should avoid this situation.