Add new_queue() method to Process, for creating user queues
In a scenario where you have multiple background process instances (like
producers and consumers), you may want to use a queue to pass data
around the instances. To ensure the queues are hooked up correctly, the
same multiprocessing context must be used. To make this easier, I added
a queue "factory" method to the Process class.