Turn docker paths to lower case
Docker image paths must be lower case and so we have to convert them to be sure.
For example, in GitLab it is possible to use a group or project name with upper
case characters and that need to be folded to lower case before we can use it.

I opted for a make internal solution, although it looks a little long and
unwieldy, as it avoids externals calls.