Add emergency-stop function
This makes it possible to stop a background worker without changing the
configuration. This is important when NSO is busy holding the transaction-lock
on something else and you immediately want to stop the background worker.

I've chosen to include the ncs Action class to actually run the emergency-stop
too, so the user only needs to wire it up to the YANG model for their
application. The tbgw test-package is effectively an example of how to do this.

We've talked about having a reusable grouping for this, which isn't included
here. I figure we can add that later on.