Add a restart action for the worker
Similar to the emergency-stop action, this acts on the worker without
configuration changes. It is useful to restart the worker when you just want to
restart the worker (disabling and enabling through configuration takes two
transactions and is just convoluted in comparison) and after an emergency-stop
when the stop failed to disable the worker in configuration, this can be useful.