Avoid race in background process restart
The way we used to implement the restart was by simply running the worker_stop()
and worker_start() function directly from the restart action callback. The
problem with that approach is that the supervisor will notice that the child is
dead after the worker_stop() function and thus start it. At the same time, we
also explicitly run worker_start(), which means we will now have two child
processes. It's a race.

It seems quite consistent on this repo, where the supervisor will start up the
child. On another repo I saw the supervisor only starting up the child
sometimes, probably because it took every slightly longer to detect that the
child was not running, thus we had time to come in and start up the background
worker, the supervisor would see it and then not start another child.

Regardless, this fixes this by instead of directly invoking what probably should
be internal methods (worker_stop() & worker_start()), we ask the supervisor to
restart the background worker process for us.

The restart test case is also expanded to cover this by ensuring there is
exactly 3 python processes.

One change is that the restart action is now asynchronous instead of blocking
and only returning after the restart is complete. That was never an intentional
part of the design so I think it's fine that the behavior changes.