Support running in different HA-modes
We have defaulted to running the background process when HA is either
disabled or when the node is in HA-mode=master.

There is now an additional argument to background_process that specifies
in what HA modes to run. It defaults to 'master' keeping the old
behavior but can be set to 'always' to always run, regardless which HA
mode we are in or 'slave' to run when a slave.

New test cases covers the new scenarios.