Refactor RestartWorker and EmergencyStop actions
When NSO 6.x+ uses the multiprocessing model for callbacks, the
ncs.application.Application.setup() method is executed in the parent
process, but the callbacks (service, action, ...) classes are started in
worker processes. This implies that any init_args passed to these
callbacks must support transfer across the process boundary = be
picklable! The restart and emergency-stop actions no longer require the
supervisor thread (which is *not* picklable) but rather use a true
multiprocessing.Queue to pass messages.